+++ foxes/models/turbine_models/set_farm_vars.py
@@ -87,11 +87,6 @@ class SetFarmVars(TurbineModel):
 
         for i, v in enumerate(self.vars):
 
-            if not isinstance(self._vdata[i], np.ndarray):
-                raise TypeError(
-                    f"Model '{self.name}': Wrong data type for variable '{v}': Expecting '{np.ndarray.__name__}', got '{type(self._vdata[i]).__name__}'"
-                )
-
             data = np.full((algo.n_states, algo.n_turbines), np.nan, dtype=FC.DTYPE)
             data[:] = self._vdata[i]