+++ foxes/output/__init__.py
@@ -4,3 +4,5 @@ from .farm_results_eval import FarmResultsEval
 from .flow_plots_2d import FlowPlots2D
 from .rose_plot import RosePlotOutput, StatesRosePlotOutput
 from .results_writer import ResultsWriter
+from .state_turbine_map import StateTurbineMap
+from .turbine_type_curves import TurbineTypeCurves