+++ foxes/utils/__init__.py
@@ -4,6 +4,7 @@ from .helpers import all_subclasses
 from .dict import Dict
 from .data_book import DataBook
 from .plotly import show_plotly_fig
+from .cubic_roots import cubic_roots
 
 from . import two_circles
 from . import abl