+++ foxes/utils/data_book.py
@@ -149,6 +149,11 @@ class DataBook:
             Flag for raising KeyError, otherwise return None,
             if context of file_name not found
 
+        Returns
+        -------
+        path : pathlib.Path
+            The path
+
         """
 
         if check_raw: